Fixed connection between Data an Web

+++ Data/Models/Equipment.cs
@@ -5,7 +5,7 @@ public class Equipment
     public int EquipmentId { get; set; }
     public string Name { get; set; }
     public string Description { get; set; }
-    public decimal PriceInGold { get; set; }
+    public double PriceInGold { get; set; }
 
     public List<Inventory> Inventories { get; set; }
     public List<PurchaseHistory> PurchaseHistories { get; set; }
